+++ src/stats.cpp
@@ -8,6 +8,7 @@
 #include "evaluator_log.hpp"
 #include "file.hpp"
 #include "log.hpp"
+#include "oeis_program.hpp"
 #include "oeis_sequence.hpp"
 #include "parser.hpp"
 #include "program_util.hpp"
@@ -403,89 +404,8 @@ void Stats::finalize() {
     blocks = blocks_collector.finalize();
   }
   if (latest_program_ids.empty()) {
-    collectLatestProgramIds();
-  }
-}
-
-void Stats::collectLatestProgramIds() {
-  auto progs_dir = Setup::getProgramsHome();
-  static const size_t max_commits = 100;            // magic number
-  static const size_t max_added_programs = 750;     // magic number
-  static const size_t max_modified_programs = 250;  // magic number
-  if (!isDir(progs_dir + ".git")) {
-    Log::get().warn(
-        "Cannot read commit history because the .git folder was not found");
-    return;
-  }
-  const std::string git_tmp = getTmpDir() + "loda_git_tmp.txt";
-  git(progs_dir, "log --oneline --format=\"%H\" -n " +
-                     std::to_string(max_commits) + " > \"" + git_tmp + "\"");
-  std::string line;
-  std::vector<std::string> commits;
-  {
-    // read commits from file
-    std::ifstream in(git_tmp);
-    while (std::getline(in, line)) {
-      if (!line.empty()) {
-        commits.push_back(line);
-      }
-    }
-  }
-  std::remove(git_tmp.c_str());
-  if (commits.empty()) {
-    Log::get().warn("Cannot read programs commit history");
-    return;
-  }
-  std::set<int64_t> ids;
-  size_t num_added_ids = 0, num_modified_ids = 0;
-  for (const auto &commit : commits) {
-    if (num_added_ids >= max_added_programs &&
-        num_modified_ids >= max_modified_programs) {
-      break;
-    }
-    git(progs_dir, "diff-tree --no-commit-id --name-status -r " + commit +
-                       " > \"" + git_tmp + "\"");
-    {
-      // read changed file names from file
-      std::ifstream in(git_tmp);
-      std::string status, path;
-      while (std::getline(in, line)) {
-        std::istringstream iss(line);
-        iss >> status;
-        iss >> path;
-        if (path.size() >= 11 && path.substr(path.size() - 4) == ".asm") {
-          auto id_str = path.substr(path.size() - 11, 7);
-          try {
-            OeisSequence seq(id_str);
-            if (isFile(seq.getProgramPath())) {
-              if (status == "A" && num_added_ids < max_added_programs) {
-                ids.insert(seq.id);
-                num_added_ids++;
-              } else if (status == "M" &&
-                         num_modified_ids < max_modified_programs) {
-                ids.insert(seq.id);
-                num_modified_ids++;
-              }
-            }
-          } catch (const std::exception &) {
-            // ignore because it is not a program of an OEIS sequence
-          }
-        }
-      }
-    }
-    std::remove(git_tmp.c_str());
-  }
-  latest_program_ids.clear();
-  for (auto id : ids) {
-    if (id >= static_cast<int64_t>(latest_program_ids.size())) {
-      const size_t new_size =
-          std::max<size_t>(id + 1, 2 * latest_program_ids.size());
-      latest_program_ids.resize(new_size);
-    }
-    latest_program_ids[id] = true;
-  }
-  if (latest_program_ids.empty()) {
-    Log::get().warn("Cannot read programs commit history");
+    latest_program_ids =
+        OeisProgram::collectLatestProgramIds(100, 750, 250);  // magic number
   }
 }